Encapsulated all hit point functionality
The functionality has been moved into the new classes: ModifyHitPoints
(class that holds the shared functionality of all DungeonCharacters),
HeroPointModifier (specific functions of heros), MonsterPointModifier
(specific functions of monsters) and HPModifierFactory (chooses which
modifier is instantiated based off of the character type being
instantiated). Tested and found one bug with refactor, Sorceress was
reporting own damage was blocked when performing attack. This was caused
by a simple logic mistake and was fixed.